+++ tests/test_pretty.py
@@ -2,7 +2,7 @@ import collections
 import io
 import sys
 from array import array
-from collections import UserDict, defaultdict
+from collections import UserDict, defaultdict, deque
 from dataclasses import dataclass, field
 from typing import Any, List, NamedTuple
 
@@ -490,6 +490,33 @@ def test_defaultdict():
     test_dict = defaultdict(int, {"foo": 2})
     result = pretty_repr(test_dict)
     assert result == "defaultdict(<class 'int'>, {'foo': 2})"
+
+
+def test_deque():
+    test_deque = deque([1, 2, 3])
+    result = pretty_repr(test_deque)
+    assert result == "deque([1, 2, 3])"
+    test_deque = deque([1, 2, 3], maxlen=None)
+    result = pretty_repr(test_deque)
+    assert result == "deque([1, 2, 3])"
+    test_deque = deque([1, 2, 3], maxlen=5)
+    result = pretty_repr(test_deque)
+    assert result == "deque([1, 2, 3], maxlen=5)"
+    test_deque = deque([1, 2, 3], maxlen=0)
+    result = pretty_repr(test_deque)
+    assert result == "deque([], maxlen=0)"
+    test_deque = deque([])
+    result = pretty_repr(test_deque)
+    assert result == "deque([])"
+    test_deque = deque([], maxlen=None)
+    result = pretty_repr(test_deque)
+    assert result == "deque([])"
+    test_deque = deque([], maxlen=5)
+    result = pretty_repr(test_deque)
+    assert result == "deque([], maxlen=5)"
+    test_deque = deque([], maxlen=0)
+    result = pretty_repr(test_deque)
+    assert result == "deque([], maxlen=0)"
 
 
 def test_array():